+++ src/main.rs
@@ -47,6 +47,9 @@ fn main() {
         dioxus_fullstack::set_server_url(backend_url);
     }
 
+    #[cfg(feature = "server")]
+    browserinfocm::backend_init().expect("failed to init backend");
+
     // In the case of only release desktop, set a window title
     #[cfg(not(feature = "server"))]
     #[cfg(all(not(debug_assertions), feature = "desktop"))]